backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-mail==0.9.1
python-dotenv==0.19.0 